added migrations for the new configurable enums
