+++ src/app/child-dev-project/aser/model/mathLevels.ts
@@ -1,0 +1,28 @@
+import { ConfigurableEnumValue } from "../../../core/configurable-enum/configurable-enum.interface";
+
+export const mathLevels: ConfigurableEnumValue[] = [
+  {
+    id: "",
+    label: "",
+  },
+  {
+    id: "nothing",
+    label: "Nothing",
+  },
+  {
+    id: "numbers1to9",
+    label: "Numbers 1-9",
+  },
+  {
+    id: "numbers10to99",
+    label: "Numbers 10-99",
+  },
+  {
+    id: "subtraction",
+    label: "Subtraction",
+  },
+  {
+    id: "division",
+    label: "Division",
+  },
+];
