+++ src/app/child-dev-project/aser/model/readingLevels.ts
@@ -1,0 +1,28 @@
+import { ConfigurableEnumValue } from "../../../core/configurable-enum/configurable-enum.interface";
+
+export const readingLevels: ConfigurableEnumValue[] = [
+  {
+    id: "",
+    label: "",
+  },
+  {
+    id: "nothing",
+    label: "Nothing",
+  },
+  {
+    id: "read_letters",
+    label: "Read Letters",
+  },
+  {
+    id: "read_words",
+    label: "Read Words",
+  },
+  {
+    id: "read_sentence",
+    label: "Read Sentence",
+  },
+  {
+    id: "read_paragraph",
+    label: "Read Paragraph",
+  },
+];
