+++ src/app/core/admin/admin/admin.component.html
@@ -21,7 +21,7 @@
   <button (click)="childrenMigrationService.migratePhotoFormat()" mat-stroked-button>
     Migrate Children to correct photo format
   </button>
-  <button (click)="configMigrationService.migrateConfig()" mat-stroked-button>
+  <button (click)="migrateConfigChanges()" mat-stroked-button>
     Migrate Config to new format
   </button>
 </p>
