+++ src/app/core/admin/admin/admin.component.ts
@@ -13,6 +13,7 @@ import { AttendanceMigrationService } from "../../../child-dev-project/attendanc
 import { NotesMigrationService } from "../../../child-dev-project/notes/notes-migration/notes-migration.service";
 import { ChildrenMigrationService } from "../../../child-dev-project/children/child-photo-service/children-migration.service";
 import { ConfigMigrationService } from "../../config/config-migration.service";
+import { ConfigurableEnumMigrationService } from "../../configurable-enum/configurable-enum-migration.service";
 
 /**
  * Admin GUI giving administrative users different options/actions.
@@ -43,7 +44,8 @@ export class AdminComponent implements OnInit {
     public attendanceMigration: AttendanceMigrationService,
     public notesMigration: NotesMigrationService,
     public childrenMigrationService: ChildrenMigrationService,
-    public configMigrationService: ConfigMigrationService
+    public configMigrationService: ConfigMigrationService,
+    public configurableEnumMigrationSerivice: ConfigurableEnumMigrationService
   ) {}
 
   ngOnInit() {
@@ -56,6 +58,11 @@ export class AdminComponent implements OnInit {
    */
   updatePhotoFilenames() {
     this.childPhotoUpdateService.updateChildrenPhotoFilenames();
+  }
+
+  async migrateConfigChanges() {
+    await this.configMigrationService.migrateConfig();
+    await this.configurableEnumMigrationSerivice.migrateSelectionsToConfigurableEnum()
   }
 
   /**
